+++ mpc_energy/MPC.py
@@ -101,7 +101,7 @@ class MPC:
         self.sell_price_uncertainty_discount_per_hour = 1    # -%/hr applied to future sell prices
         self.max_price_uncertainty_adjustment = 30           # Cap the absolute buy/sell adjustment (+/-30%)
 
-        if(self.retailer == "flow"):
+        if(self.retailer == "flow" or self.retailer == "generic"):
             self.buy_price_uncertainty_premium_per_hour = -0.01      # +%/hr applied to future buy prices (testing a negative number to encourge buying as late as possible to rely less on load forecast)
             self.sell_price_uncertainty_discount_per_hour = 0 # Flow sell prices are known with certainty
             self.max_price_uncertainty_adjustment = 1           # Cap the absolute buy/sell adjustment (+/-30%)