+++ mpc_energy/CHANGELOG.md
@@ -5,7 +5,8 @@
 * Improved Effective Price and Next Grid Interaction calculations.
 * Fixed Grid Import Mode for GoodWe Systems.
 * Fixed issue with Export Excess Solar that limited solar power.
-* Added Timed Optional Load ****
+* Added Timed Optional Load
+* 
 
 ## 0.8.0
 * **!!!Please enter the general config page to sync old config through HA with new web dashboard config!!!**.